Simplify address formatting
Added some helper functions to the AddressUtils class to simplify converting an address object into a formatted string. This not only simplifies usage, but also centralizes all interation with the google i18n library to one file so if we eventually want to swap it out, it will be very little work.

Ticket #137