Refresh notifications from the api for specific comment events
If the event is PUSH_COMMENT or DELETE_COMMENT we refresh from the api.
This is to keep the notifications db in sync. If the comment status
was changed to spam or trash, or if the comment was deleted, the
associated notification will no longer be returned from the api.