Added the proper style tag to style the review buttons per designs
We used to have this set at the theme-level of this project, but it was
oddly causing conflicts whenever we needed to tweek a style or create
a new one. For example, the white button with purple letters works fine
because that was the default style, but trying to apply a style where
the button would have a border who get overridden by the default style.