Use only the isJetpackConnected property for jetpack validation
This applies only to non-atomic stores. Due to some bugs in the
connected site info endpoint the other properties are not currently
reporting properly and this is causing the check for jetpack to fail
during the login process. All we need to know is if it's connected so
I've verified this field will work for our uses.