Move SqlScout instantiation to the Application class
This is to prevent the constant errors caused by SqlScout when testing
the app. These errors are due to the LifeCycle of the Activity causing
SqlScout to stop, destroy, and restart over and over, but putting this
code in the MainActivity is how they recommend doing it. After some
research I found several articles that recommend moving it to the
Application class to prevent this constant tear-down/rebuild error
cycle.