Fix localization issue with tab names
The localized title of the tab was being stored as the tag, and
that tag was then being used to create the filter which did not work
at all in other languages since filters are all performed in english.
To keep with the dynamic creation of tabs from the string array, I've
added some new utility methods for fetching strings by a specific
localization for these types of scenarios.