Add missing error logging where needed
We had a user report that nothing would happen after clicking either
of the buttons on the sitepicker view in the final login step and there
was nothing in the log to point to what the issue was. I found a couple
places where errors were not being logged and added logging so next time
we can better troubleshoot something like this.