Give toolbar it's own theme to prevent issues with material coloring
colorControlNormal is used by the MaterialComponents theming to style
the color of unchecked/unselected components like radios and widgets.
We want the "up" and "more" button on the toolbar to be white, but we
don't want the same for radios and checkboxes so the toolbar has its
own theme now.