Fix issue with local order date conversion and grouping
This issue was due to the local date created from the orders's UTC date
string being treated as if it were a UTC date so when calculating relative
date buckets and friendly date labels we were using the UTC version of
the device date. The confusion comes from the naming of the method
DateTimeUtils.dateUTCFromIso8601() which could be misconstrued as
returning a Date object in UTC, but really it's taking an ISO8601 UTC
datestring and parsing it into a local Date object.
The formatter used sets the timeZone as "UTC" prior to parsing out the
date object so the conversion to device timezone is done appropriately.
During the fix I found other instances where this conversion wasn't
being properly understood and fixed those as well. Since the method is
in the WordPress-Utils library, I added some comments to the DateUtils
class in this repo to make it clear that it's returning a local Date
object. I also cleaned up some lint errors.