Update password for screenshot user
I was seeing an issue while running the screenshots automation and
thought it might be due to the password. The password turned out not to
be the source of the issue, I should have tried it manually first.

Unfortunately, I can't change the password back to the previous one, so
I have to make this commit.