Cleanup code and remove extra view layer
This just fixes something I did early on but have since learned is not
necessary. Dialogs don't need to be contained in a card, the theme will
take care of the surface and elevation for us - so the extra layer
in the view hierarchy is just dead weight and a performance hit.