Fix issue with order list tabs show up after marking order complete
This was happening because the request to re-open the order detail
view was being routed through the MainActivity, so the
onReturnedFromChildFragment() method was getting called, but there was
no logic to then hid the tabs once the order detail was displayed since
it wasn't opened through the order list fragment. Added a complimentary
onChildFragmentOpened() to allow top level fragments to do any cleanup
needed.