Add screen object pattern implementation + screengrab to write UI tests
All this is based on Jeremy's work from
https://github.com/woocommerce/woocommerce-android/pull/2096

The idea behind the screen object (or page object in the web world)
patter is to build dedicated objects for each screen –or even fragment–
in the UI, each exposing methods on how to interact with it.

This way, tests can be tidier and become easier to maintain, because
for every UI change there is a self-contained and isolated location to
update.

See https://www.martinfowler.com/bliki/PageObject.html

Fastlane's screengrab is integrated in this same commit because it
seemed like too much work to untangle it from Jeremy's work for the
sake of having more step by step commits.

In order to verify the setup, a simple UI test has been added,
`ScreenshotTest.kt`. We'll build on top of it to eventually take all of
the screenshots.

Co-authored-by: Jeremy Massel <jeremy.massel@gmail.com>