Only force apply the DayNight theme once per session
Fixes #2449 Every time a new view is loaded the `getTheme()` method is
called and this was leading to a TON of overhead as the current theme
was completely overwritten and applied over and over again for every
view rendered. This was leading to OOM crashes on Android 5.1 and heavy
GC in all versions of Android.