Fix build failures in androidTest
I run a Git bisect and it identified 3463bbe58 as the breaking commit.
I'm not sure whether that's correct or not. Going by the diffs only, I
would have said 269129e49 is the culprit.

Regardless, the tests now _build_ but they don't pass. They weren't
passing in those commit either, for what it's worth.