Check for null before accessing views inside onTabSelected handler
Fixes #2419 by checking if the views in this handler are null before
attempting to perform actions on them. This is important because this
listener is being set on the tablayout, but the user currently has the
ability to switch between two different versions of this view, the
old one and the v4 version.