Rename string to fix lint error on circleci
Not sure why this is but I'm guessing it has something to do with the fact that this string was modified to contain a variable but the other translations would still not have a variable so ktlint was throwing an error that the string was not properly formatted. Changing the name of the string variable seems to have fixed this.