Refactor size function to reduce cyclomatic complexity
By replacing the 'if/else' statement with a single 'when' statement,
which adheres to an improved logic, the cyclomatic complexity of the
size function drastically decreases. This makes sure that future updates
to this function will become easier and less error prone.

Actually the idea was (kind of) stolen from the iOS equivalent solution.
Kudos to them for coming up with this elegant way of getting the size
string.