Fix product display for single product orders within order detail screen
When a user enters the order details screen that contains a single
product, then the product section was displaying 'Products' as plural
instead of 'Product' as singular.

This commit fixes this issue, which was actually a bug in the code due
to the misuse of the 'getQuantityString(...)' function. Since the 'zero'
argument is before 'one', when not using named arguments it is easy to
misuse it. By explicitly adding named arguments to select 'one' instead
of the default 'zero' parameter for the singular 'Product' string the
problem is automatically fixed.