Fix product variations list scrolling to top when load more is triggered
This change makes sure that the adapter is not set every time a new set
of product variations are fetched from the database. The current
situation was actually created by design since the ordering of product
variations on the api's side and on the database' side was different. As
such pagination could not work.

This commit fixes partially this problem. Another change was already
made on the 'FluxC' side to align the ordering of product variations on
both, the api and database, by 'date'. Both changes together will make
it possible for pagination to work as expected. But, note that the
ordering of product variations now changes, from 'menu' to 'date', which
change the product's logic for that screen.