Track ORDERS_LIST_LOADED
Note that this is slightly different to iOS. iOS tracks ORDERS_LIST_LOADED when
loading a page and loading all Order data. Android only loads the id and
dateCreated if I understand correctly.

But this is enough for our purpose. We will only use this event to get a better idea
of how fast the orders are being loaded in our customer's devices.