Fetch the latest product from the API after generating a variation
In this way, we can ensure that whatever is saved in the database is the latest.
More importantly, the `numVariations` will have the correct value.

Without this trick, the `numVariations` will either be zero or 1 after generating the
variation. Previously, we relied on the `createEmptyVariation` and the
`saveAttributeChanges` calls to _update_ the product's `numVariations`. This has
a race condition because if `createEmptyVariation` finishes last, the
`numVariations` will be zero.

What I'm not sure of is if we should make sure `saveAttributeChanges` finishes
first before generating the variation.