Fix country and state view state update on onViewCreated
With the new implementation of a search screen for country/state selection, onViewCreated is triggered when user navigates back to the address editing screen. That was causing the view state country to be replaced with empty string instead of the user selected country. This implementation checks if user selected a country (using addressDraft) and use that to update the view state. If user did not select a country in this instance of address editing screen, addressDraft.country will be empty and we can use storedAddress without problems.