Build: Update compose to 1.1.0-rc01
Also, as part of updating Jetpack Compose to '1.1.0-rc01':
- The 'kotlinCompilerVersion' Kotlin option was removed as it is
deprecated and no longer required. Jetpack Compose now defaults to using
the Kotlin compiler version defined in the build script.
- The Material Compose Theme Adapter had to be updated to its '1.1.2'
version, which also points to Jetpack Compose to '1.1.0-rc01'.

This update was done because otherwise compilation was failing due to
the below compile error:
"e: This version (1.0.5) of the Compose Compiler requires Kotlin version
1.5.31 but you appear to be using Kotlin version 1.6.0 which is not
known to be compatible.  Please fix your configuration
(or `suppressKotlinVersionCompatibilityCheck` but don't say I didn't
warn you!)."

One way of dealing with the above is to ignore it, as suggested, but
even that didn't work due to the below compile error:
"e: java.lang.NoSuchMethodError: org.jetbrains.kotlin.psi2ir.generators
.TypeTranslatorImpl.<init>(Lorg/jetbrains/kotlin/ir/util/
ReferenceSymbolTable;Lorg/jetbrains/kotlin/config/
LanguageVersionSettings;Lorg/jetbrains/kotlin/descriptors/
ModuleDescriptor;Lkotlin/jvm/functions/Function0;Lorg/jetbrains/kotlin/
ir/util/StubGeneratorExtensions;ILkotlin/jvm/internal/
DefaultConstructorMarker;)V at ..."

I also tried to update to the latest stable Jetpack Compose version, the
'1.1.0', but this didn't work as well. Doing that resulted to the below
compile error:
"e: This version (1.1.0) of the Compose Compiler requires Kotlin version
1.6.10 but you appear to be using Kotlin version 1.6.0 which is not
known to be compatible.  Please fix your configuration
(or `suppressKotlinVersionCompatibilityCheck` but don't say I didn't
warn you!)."

As I was getting the above, I then tried to update to the latest stable
Kotlin version, the '1.6.10', which also matches with the latest stable
Jetpack Compose version, the '1.1.0', but this didn't work as well.
Doing that resulted to the below compile error:
"e: This version (1.1.0) of the Compose Compiler requires Kotlin version
1.6.10 but you appear to be using Kotlin version 1.6.0 which is not
known to be compatible.  Please fix your configuration
(or `suppressKotlinVersionCompatibilityCheck` but don't say I didn't
warn you!)."

As such, and since I didn't want increase the scope of this Gradle and
AGP upgrade, I chose to temporary update Jetpack Compose to the working
'1.1.0-rc01' version, which is compatible with the Kotlin '1.6.0'
version and as such compatible with the current state of this project
and its correlated dependencies.