Build: Force kotlin version by strict kotlin compiler classpath version
This is done due to the 'Navigation Safe Args' plugin, which as of
version '2.4.1' point to Kotlin '1.6.0' instead. A future '2.5.0'
version of the 'Navigation Safe Args' plugin will point to Kotlin
'1.6.10', but that version might come way later than needed.

If Kotlin '1.6.10' is not forced the below compilation error occurs:

'This version (1.1.1) of the Compose Compiler requires Kotlin version
1.6.10 but you appear to be using Kotlin version 1.6.0 which is not
known to be compatible.  Please fix your configuration
(or `suppressKotlinVersionCompatibilityCheck` but don't say I didn't
warn you!).'

This has been also extensively discussed in this GitHub PR:
https://github.com/woocommerce/woocommerce-android/pull/5755
#issuecomment-1047165088