Machine re-indent `Fastfile`
I'm doing this in a dedicated commit, which will then be a dedicated PR,
because I'm about to work with the file as part of the localization
tooling update project and I don't want to introduce unnecessary noise
while doing that.

I decided to do this instead of a more thorough Rubocop setup and manual
indentation tweaking (e.g. by adding a new line after a `(` so the
indentation is consistent across method calls, regardless of how long
the method names are) because of time constraints on the project's
delivery.

I've done the reindent via Vim's `gg=G` (`gg` go to the
top of the file; `=G` indent `=` to the bottom of the file `G`). I'm
writing this down in case something looks odd, so we can trace back.