Use Fastlane API to commit in `localize_libs`
Notice with this change we lose the option to opt-out from committing.
As far as I know, this is something we weren't currently using, and it's
not something mentioned in the release management scenario. So, I think
it's okay to drop it instead of reproducing it in the new version.

For what is worth, WordPress Android doesn't do this either.
https://github.com/wordpress-mobile/WordPress-Android/blob/6b2e279364d76077c9c96dbc9a15dfb615982d3f/fastlane/lanes/localization.rb#L363