Analysis: Enable all warnings as errors for quick login module
This 'allWarningsAsErrors' configuration is currently applied on the
module level in order to make sure that, as the overall 'Kotlin Warnings
as Errors' work is progressing, no new warnings are added to this
module, which is already free of warnings.

When the overall 'Kotlin Warnings as Errors' work is complete on all
modules, then this module level 'allWarningsAsErrors' configuration will
be replaced by a root level 'allWarningsAsErrors' configuration that
will be applied by default to all modules from there after.