Analysis: Resolve global scope delicate coroutines api warning
1st Warning Message: "This is a delicate API and its use requires care.
Make sure you fully read and understand documentation of the declaration
that is marked as a delicate API."

Adding the '@OptIn(DelicateCoroutinesApi::class)' annotation for the
'cancelPayment(...)' function is the recommended action to resolve this
'GlobalScope' warnings. However, this results to the below '@OptIn'
warnings:

2nd Warning Message: "This annotation should be used with the
compiler argument '-Xopt-in=kotlin.RequiresOptIn'"

Adding the '-Xopt-in=kotlin.RequiresOptIn' free compiler argument is the
recommended action to resolve this '@OptIn' warning.