Analysis: Change direct experimental coroutines api with opt-in
Changing the '@ExperimentalCoroutinesApi' to an
'@OptIn(ExperimentalCoroutinesApi::class)' or an
'@OptIn(FlowPreview::class)' and placing it closer to usage will make it
more obvious which Coroutines API is actually experimental and why.

Also, this will resolve other such
'@OptIn(ExperimentalCoroutinesApi::class)' or
'@OptIn(FlowPreview::class)' warnings, which was based on the class as
it was marked directly as '@ExperimentalCoroutinesApi' instead of an
opt-in.