Analysis: Suppress progress dialog deprecated warnings
Warning Messages:
- "'ProgressDialog' is deprecated. Deprecated in Java"
- "'show(Context!, CharSequence!, CharSequence!, Boolean):
ProgressDialog!' is deprecated. Deprecated in Java"

These deprecated warnings are suppressed instead of being resolved since
a resolution would require a proper solution.

You will also notice that I suppressed the deprecated warning both, on
the file level and closer to the deprecation itself. In addition to
that, you will notice that I suppressed the deprecated warning for other
such warnings, like 'startActivityForResult(...)' and
'onActivityResult(...)'. I did that because suppressing on the file
level was necessary due to the 'android.app.ProgressDialog' import that
can't be suppressed any other way. But, when doing so, every deprecated
warning is suppressed as well. At which point the deprecated information
is lost. This way, adding first the '@Suppress("DEPRECATION")' closer to
source and then adding it on the file level makes  sure that this
information is not lost, at least for the existing code.

Explanation: "This class was deprecated in API level 26. ProgressDialog
is a modal dialog, which prevents the user from interacting with the
app. Instead of using this class, you should use a progress indicator
like ProgressBar, which can be embedded in your app's UI. Alternatively,
you can use a notification to inform the user of the task's progress."

Docs: https://developer.android.com/reference/android/app/ProgressDialog