Analysis: Suppress fluxc related update order status deprecated warnings
Warning Messages: "'UPDATE_ORDER_STATUS' is deprecated. Deprecated in
Java"

These deprecated warnings are suppressed instead of being resolved since
a resolution would require a proper solution.

You will also notice that I suppressed the deprecated warning both, on
the file level and closer to the deprecation itself. In addition to
that, you will notice that I suppressed the deprecated warning for other
such warnings, like 'SelectedSiteChangedEvent'. I did that because
suppressing on the file level was necessary due to the
'org.wordpress.android.fluxc.action.WCOrderAction.UPDATE_ORDER_STATUS'
import that can't be suppressed any other way. But, when doing so, every
deprecated warning is suppressed as well. At which point the deprecated
information is lost. This way, adding first the '@Suppress
("DEPRECATION")' closer to source and then adding it on the file level
makes sure that this information is not lost, at least for the existing
code.

Explanation:
- "Use suspendable WCOrderStore.updateOrderStatus(..) directly."
- "Event bus is considered deprecated."

Docs:
- https://github.com/wordpress-mobile/WordPress-FluxC-Android/wiki/
%5BDeprecated%5D-Architecture
- https://github.com/wordpress-mobile/WordPress-FluxC-Android/wiki/
Architecture