Analysis: Suppress fragment pager adapter deprecated warning
Warning Messages: "'FragmentPagerAdapter' is deprecated. Deprecated in
Java"

This deprecated warning is suppressed instead of being resolved since
a resolution would require a proper solution.

You will also notice that I suppressed the deprecated warning both, on
the file level and closer to the deprecation itself. I did that because
suppressing on the file level was necessary due to the
'androidx.fragment.app.FragmentPagerAdapter' import that can't be
suppressed any other way. But, when doing so, every deprecated warning
is suppressed as well. At which point the deprecated information is
lost. This way, adding first the '@Suppress("DEPRECATION")' closer to
source and then adding it on the file level makes sure that this
information is not lost, at least for the existing code.

Explanation: "This class is deprecated. Switch to
androidx.viewpager2.widget.ViewPager2 and use
androidx.viewpager2.adapter.FragmentStateAdapter instead."

Docs: https://developer.android.com/reference/androidx/fragment/app/
FragmentPagerAdapter