Analysis: Resolve non exhaustive when statements when warnings
Warning Message: "Non exhaustive 'when' statements"

Adding an else branch or remaining branches is the recommended
action to resolve this kind of when warnings.

1) You will notice that I added remain branches wherever possible, while
only added an else branch to the 'MainPresenter.onOrderChanged(...)' and
'PrivacySettingsPresenter.onAccountChanged(...)' when statements because
otherwise those will be blown up with cases that are meant to be ignored
anyway.
2) Also, you will notice that I optimized the imports on all those when
statements and the imports they were using because one or no import is
always better than having multiple imports for no particular reason.
Also, having the extra information like 'RequestResult.SUCCESS' instead
of plain 'SUCCESS' is usually better for readability purposes as the
reader will immediately understand that this 'SUCCESS' is
'RequestResult' related, without the need to navigate to the 'SUCCESS'
in order to understand when that is coming from.