Analysis: Resolve experimental coroutine api warnings for tests
Warning Message: "This declaration needs opt-in. Its usage should be
marked with '@kotlinx.coroutines.ExperimentalCoroutinesApi' or '@OptIn
(kotlinx.coroutines.ExperimentalCoroutinesApi::class'"

Adding the 'ExperimentalCoroutinesApi' annotation where
needed and at the test suite class level is the recommended action to
resolve these 'ExperimentalCoroutinesApi' warnings so that to end up
being consistent with the rest of the test related codebase.