Analysis: Resolve (no) wildcard imports detekt warnings for kotlin test
With this change all the 'kotlin.test.*' wildcard imports are being
replaced with explicit non wildcard imports as per this project
standards.

Also, with this change 1 entry is removed from the Detekt related
'baseline.xml' file as it is no longer needed. The imports were
optimized on all those files that were violating the
'kotlin.test.*' (no) wildcard imports rule.