Extract 'com.woocommerce.android' too `APP_PACKAGE_NAME` in `Fastfile`
I noticed this while working on the implementation for using
`upload_to_play_store` (`supply`) to upload the release notes and other
metadata to the Play Store.

Admittedly, we could have lived with the package name being hard coded,
as it's not likely to change anytime soon (or ever?). Still, it's useful
to have it centralized. I wasn't sure what value to use when writing my
code and it would have been handy to have a constant at hand.