Refactor:
The incorrect WPCom email notification is shown after user enters a wpcom email address in `LoginEmailFragment`.

There are two ways to enter `LoginEmailFragment`:

1. Enter site address -> Enter WPCom email
2. Enter WPCom Account.

In the second case, no site address is available. So in this case, we want to show the default login support notification, instead of the custom notification that leads to login with site credentials.