Analysis: Fix multiple subs specified in non-positional format warnings
Warning Message: "Multiple substitutions specified in non-positional
format of string resource string/xyz. Did you mean to add the
formatted="false" attribute?"

Replacing the non-positional '%s' or '%d' with the positional '%1$s' or
'%1$d' (where '1' is a non zero number), in cases with multiple '%s'
or/and '%d' per string is found, resovles this kind of string format
warnings.