Analysis: Suppress unused attribute lint warn for data extraction rules
Warning Message: "Attribute dataExtractionRules is only used in
API level 31 and higher (current min is 24)"

This attribute was added as part of
937b192aeae56af77e17deacdcbf9c314763cde0. As per the commit message
description it was meant to be used for Android 12 and above (API 31).
As such, this attribute is unused for Android 11 (API 30) and below, but
can be also suppressed, since this is expected.

Commit Message: "Introduce a new config file for backup for Android 12"