Analysis: Suppress unused attribute lint warns for order creation
Warning Message: "Attribute accessibilityHeading is only used in
API level 28 and higher (current min is 24)"

This attribute was added as part of
1b760468b38e14f7c48f98964949f455338b6a72. As per the PR description it
was meant to be used for Android 9 and above (API 28).
As such, this attribute is unused for Android 8 (API 27) and below, but
can be also suppressed, since this is expected.

PR Description: "Adding android:accessibilityHeading="true" to order
creation form headers in order to enable heading navigation. Note that
this is only supported on API 28+ but the app runs fine on older APIs."