Refactor: Replace build version code with system version utils checks
This is done in order to make sure that every class is using the
'SystemVersionUtils' utility class to check on build version codes.
By doing so it is made sure that the codebase is being consistent in
such checks, especially naming wise, as it now becomes more readable.

Also, with this change, and as long as the 'SystemVersionUtils' utility
class continues to be utilized, fixing the 'ObsoleteSdkInt' Lint
warnings, at least on the class level, will become just this bit easier
from here onwards.

Note that with that change only the below files are now accessing the
'android.os.Build' import:
- SystemVersionUtils
- DeviceInfo
- DataContainerExt