Analysis: Suppress android hardware camera deprecated warnings
Warning Messages:
- "'Camera' is deprecated. Deprecated in Java"
- "'getter for parameters: Camera.Parameters!' is deprecated.
Deprecated in Java"
- "'getter for supportedPreviewSizes: (Mutable)List<Camera.Size!>!'
is deprecated. Deprecated in Java"
- "'getter for supportedPictureSizes: (Mutable)List<Camera.Size!>!'
is deprecated. Deprecated in Java"
- "'width: Int' is deprecated. Deprecated in Java"
- "'height: Int' is deprecated. Deprecated in Java"
- "'Size' is deprecated. Deprecated in Java"
- "'CameraInfo' is deprecated. Deprecated in Java"
- "'Parameters' is deprecated. Deprecated in Java"

This deprecated warning is suppressed instead of being resolved since a
resolution would require a proper solution.

Explanation: "This class was deprecated in API level 21.
We recommend using the new android.hardware.camera2 API for new
applications."

For more info see:
- android.hardware.Camera (Docs):
https://developer.android.com/reference/android/hardware/Camera