7733 Use correct build_type key, and reorder params
We changed the definition of the key after this code was written to use `build_type`.

Additionally, we found that there’s a backend bug which means that query params after the first are ignored. As a workaround, this moves them in to priority order, so that test JITMs can be viewed, and in production, we can filter by platform.

Once the backend bug is resolved, any of the params can be used to target JITMs.