Build: Remove fetchstyle plugin from project
This 'style-config-android' plugin is mostly unused and very outdated.

To the Apps Infra's knowledge and checking GE stats no-one is using the
'./gradlew downloadConfigs' task to fetch/update the style config files.
For years now, almost all Android engineers depend on the default AS
style and such like '.idea' configuration.

------------------------------------------------------------------------

In addition to that, this plugin has recently created an build related
issue where it couldn't be found in any of the configured sources (see
below). Keeping and maintaining such an unused/outdated plugin is
counter productive.

"* What went wrong:
 [2022-11-08T07:04:47Z] Plugin [id: 'com.automattic.android.fetchstyle',
 artifact: 'com.automattic.android:fetchstyle:1.1'] was not found in any
 of the following sources:"

------------------------------------------------------------------------

As such, this plugin is now removed and will be (most probably) also
removed on all other repos (sooner or later). Then, the plugin's repo
will be then marked as deprecated and archived.

Plugin Link: https://github.com/Automattic/style-config-android