Update `update_appstore_strings` documentation to YARD-style
See conversation at
https://github.com/woocommerce/woocommerce-ios/pull/8160#discussion_r1027081798

This is the documentation format we should have going forward, and can
be seen rolled out more widely in WordPress iOS, e.g.:
https://github.com/wordpress-mobile/WordPress-iOS/blob/784c2a4f2ad74deb36da0a20d84b6fdd16ed56b6/fastlane/lanes/locali>

Updating the entire `Fastfile` is out of scope for the piece of work
this change belongs to. Hopefully this can be the first in an organic
migration process, which we could then complete once critical mass has
been reached with a focused burst.