Fix: Make webview login retrying work if initially app pwd is disabled, and then enabled before retrying.
This is to fix this bug report: https://github.com/woocommerce/woocommerce-android/pull/8618#issuecomment-1489306887

The `State.RetryWebAuthorization` state is needed to ensure the `state` emits a new value, thus triggering a new webview reload. We can't re-use `State.WebAuthorization` in this retrying case because the state is already `WebAuthorization` at this point.