Add: JetpackStatus return value if the fetch returns 403.
A 403 means Jetpack is activated (otherwise the return would be 404 as it is unable to create a response to the API call). However, in this case it also means the user does not have enough user privilege to get the connection data.

Later in the view model, we use the fact fact above to assume that the user does not have privilege to connect, and show the "admin access required" screen.