Add: Handling Jetpack connection case using application password.
The main motivation here is to allow Jetpack connection in Application Passwords login to completely avoid using cookie+nonce authentication.

Depending on the current state of Jetpack connection, there are two possibilities:

1. Jetpack is already connected at site-connection level. In this case, the Jetpack connection URL will be something that is usable right away without needing cookie+nonce authentication.

2. Jetpack is not yet connected. In this case, the Jetpack connection URL looks like https://{site_url}/wp-admin/admin.php?page=jetpack&action=register&_wpnonce={nonce} . For this connection URL to work, it requires the same authentication Cookie that's used to fetch the connection URL (more details on https://github.com/woocommerce/woocommerce-android/issues/7525 )

In the case of Application Passwords, we don't want to use authentication cookie, so this URL is unusable. Thus, as a workaround, we use an URL in the format of https://wordpress.com/jetpack/connect?url={site_url}&mobile_redirect=woocommerce://jetpack-connected&from=mobile instead. When shown on a webview, it presents merchants a way to connect. When completed,the webview will redirect to `woocommerce://jetpack-connected`, where it will quit and the app can proceed to confirm that connection is finished.

More on this workaround is available on pe5sF9-1le-p2#comment-1942