Deps: Update fluxc version to pr hash (v2)
This 'FluxC' PR hash updates the library to that branch version
where 'Gradle & AGP' is also upgraded to version '8.2.1 & 8.0.2'
respectively.

FluxC PR: https://github.com/wordpress-mobile/
WordPress-FluxC-Android/pull/2772

This step is required in order to check that these 'FluxC'
related changes work as expected for WCAndroid.

------------------------------------------------------------------------

This is done to make sure that the new fluxc pr hash is pointing to the
most recent 'trunk' version of 'FluxC'.