Release: Remove version codes to retain config for 356 (v10.9.2)
This retained version was used to have an APK uploaded on PlayStore
that will keep supporting Android 5/6 (minSdkVersion 21) user, which
would be able to download at least this older version of the app to.

However, and as per this recent 'pdnsEh-1gN-p2' post, due to Google
starting to enforce the 'targetSdkVersion' requirement, it has been
decided that to no longer support Android 5/6 users in order avoid
risking being rejected by Google on our main APK and production track.