Add a fixed version of hiltNavGraphViewModels extension
The default implementation from the androidx library doesn't supply the `extrasProducer` argument to the `createViewModelLazy` function, this breaks the state restoration when the scoped ViewModel is recreated after a process death from a child Fragment, as its SavedState will be linked to the Fragment itself and not the NavBackStackEntry.

The fix is to supply the `extrasProducer` argument to the `createViewModelLazy` function.