Build: Update fluxc version to pr hash (#2886)
This 'FluxC' PR hash updates the library to that branch version
where the 'MediaModel' is updated to its new null proof version.

FluxC PR: https://github.com/wordpress-mobile/
WordPress-FluxC-Android/pull/2886

This step is required in order to check that these 'FluxC'
related changes work as expected for WCAndroid.

------------------------------------------------------------------------

FYI: This change is an addition to
faf148bae26637eff20f4732f792b277472d1170. This change is necessary in
order to overcome and resolve new merge conflicts.