Restore behavior of tests passing despite exceptions thrown
This is a hacky workaround for the change in kotlinx.coroutines 1.7.0 that causes tests that throw exceptions to fail. Previously test methods that threw exceptions would not prevent tests from passing, which was a bug in kotlinx.coroutines that has now been fixed. However, significant number of our tests are currently failing because of this change, because we have bugs in the implementation of our test classes, usually related to the way we instantiate the view models in tests.

See the following issue for more details: https://github.com/Kotlin/kotlinx.coroutines/issues/1205. The workaround coming with this commit is taken from the related PR: https://github.com/Kotlin/kotlinx.coroutines/pull/3736 and is a solution suggested by JetBrains to disable the new behavior using non-public API until we fix our tests. This should not be considered a long-term solution, rather a temporary hack.