CI: Change dep tree diff config from vanilla release to jalapeno debug
This is because the 'cp gradle.properties-example gradle.properties'
command is also copying the default ignored variants configuration,
which exclude all the 'vanilla' and 'wasabi' product flavors, as well as
the 'release' build types. This leaves only the 'jalapeno' product
flavor and 'debug' build type as active variants. Otherwise the build,
and as such this job will fail in CI.