Update `release-toolkit` to latest 9.x to fix QR code in GitHub
The fix was in 9.3.1, but it doesn't hurt to get the latest 9.x
available.

For an example of broken QR code, see
https://github.com/woocommerce/woocommerce-android/pull/11670

Also notice that `release-toolkit` is now at version 11.x. A major
version upgrade of that kind is out of scope for me at the moment. I
just wanted to get this useful quick win.