Move responsibility of cart's opacity animating to cart
It's the Cart that keeps in its internal state the selected items count or empty state. That's why the Cart should decide about the opacity of its overlay and control the animation. Home is not keeping that state and is not capable of knowing if the Cart is really empty or not.