Resolve `androidx.navigation.safeargs.kotlin` in main module
This way, `androidx.navigation.safeargs.kotlin` applied in `WooCommerce` module isn't resolved in isolation, causing creating another classpath.

In this particular example, applying this plugin in `WooCommerce` caused `protobuf-java` in version `3.17.2` to be applied. This dependency is already on classpath on main module in version `3.22.3` so resolving `safeargs` in root module classpath will bump `protobuf` to higher version.