CI: Add install gmd image and accept android licence before assemble
In order to use Gradle Managed Devices (GMD) on CI for a particular
device, it needs to be installed and its license accepted.

------------------------------------------------------------------------

FYI: After trying that, and without deeper investigation as to why, this
attempt fails on CI with the below:

[2024-08-28T08:39:58Z] [=== ] 10% Installing Google APIs Intel x86_6
Warning: Failed to read or create install properties file.
[2024-08-28T08:39:58Z] 🚨 Error: The command exited with status 1
[2024-08-28T08:39:58Z] user command error: exit status 1

It seems that the installation process does not have the necessary
permissions to write to the directory where the SDK is located, thus
failing.