Deps: Migrate androidx core to version catalogs on all modules
FYI: Notice that the 'AndroidX Core Splashscreen' version on the
'WooCommerce' module is pointing to '1.0.0', while on the
'WooCommerce-Wear' module it is pointing to '1.0.1'. This change
also updates the 'WooCommerce' version to point to '1.0.1' as well, it
being just a batch update that includes a single bugfix and no more.

Release Notes: https://developer.android.com/jetpack/androidx/releases/
core#core-splashscreen-1.0.1