Deps: Migrate androidx preference to version catalogs on all modules
FYI: Notice that the 'AndroidX Preference KTX' version on the
'WooCommerce' module is pointing to '1.2.0', while on the
'WooCommerce-Wear' module it is pointing to '1.2.1'. This change
also updates the 'WooCommerce' version to point to '1.2.1' as well, it
being just a batch update that includes a few bugfixes and no more.

Release Notes: https://developer.android.com/jetpack/androidx/releases/
preference#1.2.1