Deps: Migrate androidx datastore to version catalogs on all modules
FYI: Notice that the 'AndroidX Datastore' versions on the
'WooCommerce' module are pointing to '1.0.0', while on the
'WooCommerce-Wear' module they are pointing to '1.1.0'. This change
also updates the 'WooCommerce' versions to point to '1.1.0' as well, it
having only additional features on top of '1.0.0' but no breaking
changes.

Release Notes: https://developer.android.com/jetpack/androidx/releases/
datastore#1.1.0