Deps: Migrate androidx hilt to version catalogs on all modules
FYI: Notice that the 'AndroidX Hilt Compose' version on the
'WooCommerce-Wear' module is pointing to '1.2.0', while on the
'WooCommerce' module it is pointing to '1.1.0', as well as all other
'AndroidX Hilt' dependencies on 'WooCommerce-Wear' too, which too are
pointing to '1.1.0'. This change downgrades the 'WooCommerce-Wear'
version of only 'AndroidX Hilt Compose' to point to '1.1.0' as well,
'1.2.0' has only one important change on top of '1.1.0' but no breaking
changes.

Release Notes: https://developer.android.com/jetpack/androidx/releases/
hilt#hilt_version_12_2